backend/pyproject.toml
[project]
name = "heros-server"
version = "0.1.1"
description = "HEROSLAB"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings>=2.7.0",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
    "psycopg[binary,pool]>=3.2.3",
    "requests>=2.32.3",
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "dateutils>=0.6.12",
    "typst>=0.12.3",
    "python-multipart>=0.0.20",
]

[project.scripts]
heros-server = "heros.server:main_server"

[tool.uv]
package = true

[tool.ruff]
line-length = 100
indent-width = 4


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B"]
ignore = ["E501"]

[dependency-groups]
dev = [
    "pyright>=1.1.391",
    "python-language-server[pycodestyle,rope]>=0.36.2",
    "pytest>=8.3.4",
    "setuptools>=75.6.0",
]

[tool.setuptools.packages.find]
include = ["heros"]
